+++ src/directives/_class.js
@@ -1,0 +1,6 @@
+export function _class(element, value) {
+  if (!isObject(value)) return console.error("error class");
+  Object.entries(value).forEach(([className, condition]) => {
+    element.classList.toggle(className, !!condition);
+  });
+}
