+++ src/directives/index.js
@@ -1,0 +1,7 @@
+import { _class } from './_class.js'
+import { _text } from './_text.js'
+
+export default {
+	_class,
+	_text
+}
